internal/contour: move annotation processing to its own file
